+++ packages/daisyui/test/helpers/createMocks.ts
@@ -6,8 +6,8 @@ import Templates from '../../src/templates/Templates';
 import generateWidgets from '../../src/widgets/Widgets';
 
 // Mock any problematic dependencies
-// This helps when a dependency like nanoid uses ESM
-jest.mock('nanoid', () => ({
+// This helps when a dependency like uniqueId uses ESM
+jest.mock('lodash/uniqueId', () => ({
   nanoid: () => 'test-id-12345',
 }));